chore(tools): extract apply patch engine
