+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "crates/rara-mcp-client",
     "crates/rara-state",
     "crates/rara-memory",
+    "crates/rara-apply-patch",
     "crates/rara-observability",
     "crates/rara-tools",
     "crates/rara-plugins",
