+++ crates/rara-apply-patch/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rara-apply-patch"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+thiserror = "2"
